alteracao int  calculo valor total data.py linha 547

+++ engine/data.py
@@ -542,9 +542,9 @@ ENGINE=InnoDB;''')
                     calculo_minuto = (2/4) 
                 elif minutos_passados > 15 and minutos_passados <= 30:
                     calculo_minuto = (3/4) 
-                valor_total = ((dias_passados * 24 * 60) * 50)
-                valor_total = valor_total + (calculo_minuto * taxa )
-                valor_total = valor_total + calculo_hora * 15
+                valor_total = int((dias_passados * 24 * 60) * 50)
+                valor_total = int(valor_total + (calculo_minuto * taxa ))
+                valor_total = int(valor_total + calculo_hora * 15)
                 result = self.cobranca_excedente(dias_passados, calculo_hora, calculo_minuto, self.__locacao[0][0])#(valor_total,hj)
